Revert "	new file:   AudioPilot.py"
