--- Data.py
@@ -1,341 +1,0 @@
-# offest for dB values
-gainOffest = 38
-
-# hardcoded frequencies based on /meters/15 data
-frequencies = [
-    20, 21, 22, 24, 26, 28, 30, 32, 34, 36,
-    39, 42, 45, 48, 52, 55, 59, 63, 68, 73,
-    78, 84, 90, 96, 103, 110, 118, 127, 136, 146,
-    156, 167, 179, 192, 206, 221, 237, 254, 272, 292,
-    313, 335, 359, 385, 412, 442, 474, 508, 544, 583,
-    625, 670, 718, 769, 825, 884, 947, 1020, 1090, 1170,
-    1250, 1340, 1440, 1540, 1650, 1770, 1890, 2030, 2180, 2330,
-    2500, 2680, 2870, 3080, 3300, 3540, 3790, 4060, 4350, 4670,
-    5000, 5360, 5740, 6160, 6600, 7070, 7580, 8120, 8710, 9330,
-    10000, 10720, 11490, 12310, 13200, 14140, 15160, 16250, 17410, 18660
-]
-
-# Initialize dataRTA with all frequencies set to a default list with a placeholder dB value
-dataRTA = {freq: [-90] for freq in frequencies}
-
-# Define band ranges
-bandRanges = { # band_name: [(freq_id, freq), ...]
-#    'LowCut': [(0.2600, 120.5)],
-    'Low': [(0.2650, 124.7),        # 124.7 to 306.2
-        (0.2700, 129.1),
-        (0.2750, 133.7),
-        (0.2800, 138.4),
-        (0.2850, 143.2),
-        (0.2900, 148.3),
-        (0.2950, 153.5),
-        (0.3000, 158.9),
-        (0.3050, 164.4),
-        (0.3100, 170.2),
-        (0.3150, 176.2),
-        (0.3200, 182.4),
-        (0.3250, 188.8),
-        (0.3300, 195.4),
-        (0.3350, 202.3),
-        (0.3400, 209.4),
-        (0.3450, 216.8),
-        (0.3500, 224.4),
-        (0.3550, 232.3),
-        (0.3600, 240.5),
-        (0.3650, 248.9),
-        (0.3700, 257.6),
-        (0.3750, 266.7),
-        (0.3800, 276.1),
-        (0.3850, 285.8),
-        (0.3900, 295.8),
-        (0.3950, 306.2)],
-    'Low Mid': [(0.4000, 317.0),    # 317.0 to 1055.0
-        (0.4050, 328.1),
-        (0.4100, 339.6),
-        (0.4150, 351.6),
-        (0.4200, 363.9),
-        (0.4250, 376.7),
-        (0.4300, 390.0),
-        (0.4350, 403.7),
-        (0.4400, 417.9),
-        (0.4450, 432.5),
-        (0.4500, 447.7),
-        (0.4550, 463.5),
-        (0.4600, 479.8),
-        (0.4650, 496.6),
-        (0.4700, 514.1),
-        (0.4750, 532.1),
-        (0.4800, 550.8),
-        (0.4850, 570.2),
-        (0.4900, 590.2),
-        (0.4950, 611.0),
-        (0.5000, 632.5),
-        (0.5050, 654.7),
-        (0.5100, 677.7),
-        (0.5150, 701.5),
-        (0.5200, 726.2),
-        (0.5250, 751.7),
-        (0.5300, 778.1),
-        (0.5350, 805.4),
-        (0.5400, 833.7),
-        (0.5450, 863.0),
-        (0.5500, 893.4),
-        (0.5550, 924.8),
-        (0.5600, 957.3),
-        (0.5650, 990.9),
-        (0.5700, 1002.0),
-        (0.5750, 1006.0),
-        (0.5800, 1009.0),
-        (0.5850, 1013.0),
-        (0.5900, 1017.0),
-        (0.5950, 1021.0),
-        (0.6000, 1026.0),
-        (0.6050, 1030.0),
-        (0.6100, 1035.0),
-        (0.6150, 1039.0),
-        (0.6200, 1044.0),
-        (0.6250, 1049.0),
-        (0.6300, 1055.0)],
-    'High Mid': [(0.6350, 1600),    # 1600 to 5020
-        (0.6400, 1660),
-        (0.6450, 1720),
-        (0.6500, 1780),
-        (0.6550, 1840),
-        (0.6600, 1910),
-        (0.6650, 1970),
-        (0.6700, 2040),
-        (0.6750, 2110),
-        (0.6800, 2190),
-        (0.6850, 2270),
-        (0.6900, 2340),
-        (0.6950, 2430),
-        (0.7000, 2510),
-        (0.7050, 2600),
-        (0.7100, 2690),
-        (0.7150, 2790),
-        (0.7200, 2890),
-        (0.7250, 2990),
-        (0.7300, 3090),
-        (0.7350, 3200),
-        (0.7400, 3310),
-        (0.7450, 3430),
-        (0.7500, 3550),
-        (0.7550, 3680),
-        (0.7600, 3810),
-        (0.7650, 3940),
-        (0.7700, 4080),
-        (0.7750, 4220),
-        (0.7800, 4370),
-        (0.7850, 4520),
-        (0.7900, 4680),
-        (0.7950, 4850),
-        (0.8000, 5020)],
-    'High': [(0.8050, 5200),        # 5200 to 20000
-        (0.8100, 5380),
-        (0.8150, 5570),
-        (0.8200, 5760),
-        (0.8250, 5970),
-        (0.8300, 6180),
-        (0.8350, 6390),
-        (0.8400, 6620),
-        (0.8450, 6850),
-        (0.8500, 7090),
-        (0.8550, 7340),
-        (0.8600, 7600),
-        (0.8650, 7870),
-        (0.8700, 8140),
-        (0.8750, 8430),
-        (0.8800, 8730),
-        (0.8850, 9030),
-        (0.8900, 9350),
-        (0.8950, 9680),
-        (0.9000, 10020),
-        (0.9050, 10370),
-        (0.9100, 10740),
-        (0.9150, 11110),
-        (0.9200, 11500),
-        (0.9250, 11890),
-        (0.9300, 12330),
-        (0.9350, 12760),
-        (0.9400, 13210),
-        (0.9450, 13670),
-        (0.9500, 14150),
-        (0.9550, 14650),
-        (0.9600, 15170),
-        (0.9650, 15700),
-        (0.9700, 16250),
-        (0.9750, 16820),
-        (0.9800, 17410),
-        (0.9850, 18030),
-        (0.9900, 18660),
-        (0.9950, 19320),
-        (1.0000, 20000)]
-}
-
-qValues = {        # q_value:q_id      6.4 to 1.0
-    6.4: 0.1268, 
-    6.1: 0.1408, 
-    5.8: 0.1549, 
-    5.5: 0.1690,
-    5.3: 0.1831,
-    5.0: 0.1972,
-    4.8: 0.2113,
-    4.5: 0.2254,
-    4.3: 0.2394,
-    4.1: 0.2535,
-    3.9: 0.2676,
-    3.7: 0.2817,
-    3.5: 0.2958,
-    3.4: 0.3099,
-    3.2: 0.3239,
-    3.1: 0.3380,
-    2.9: 0.3521,
-    2.8: 0.3662,
-    2.6: 0.3803,
-    2.5: 0.3944,
-    2.4: 0.4085,
-    2.3: 0.4225,
-    2.2: 0.4366,
-    2.1: 0.4507,
-    2.0: 0.4648,
-    1.9: 0.4789,
-    1.8: 0.4930,
-    1.7: 0.5070,
-    1.6: 0.5211,
-    1.5: 0.5352,
-    1.4: 0.5634,
-    1.3: 0.5775,
-    1.2: 0.6056,
-    1.1: 0.6338,
-    1.0: 0.6479
-}
-
-eqGainValues = {    # gain_value:gain_hexa                  -15.0 to 15.0
-    -15.00: 'C1700000',
-    -14.75: 'C16C0000',
-    -14.50: 'C1680000',
-    -14.25: 'C1640000',
-    -14.00: 'C1600000',
-    -13.75: 'C15C0000',
-    -13.50: 'C1580000',
-    -13.25: 'C1540000',
-    -13.00: 'C1500000',
-    -12.75: 'C14C0000',
-    -12.50: 'C1480000',
-    -12.25: 'C1440000',
-    -12.00: 'C1400000',
-    -11.75: 'C13C0000',
-    -11.50: 'C1380000',
-    -11.25: 'C1340000',
-    -11.00: 'C1300000',
-    -10.75: 'C12C0000',
-    -10.50: 'C1280000',
-    -10.25: 'C1240000',
-    -10.00: 'C1200000',
-    -9.75: 'C11C0000',
-    -9.50: 'C1180000',
-    -9.25: 'C1140000',
-    -9.00: 'C1100000',
-    -8.75: 'C10C0000',
-    -8.50: 'C1080000',
-    -8.25: 'C1040000',
-    -8.00: 'C1000000',
-    -7.75: 'C0F80000',
-    -7.50: 'C0F00000',
-    -7.25: 'C0E80000',
-    -7.00: 'C0E00000',
-    -6.75: 'C0D80000',
-    -6.50: 'C0D00000',
-    -6.25: 'C0C80000',
-    -6.00: 'C0C00000',
-    -5.75: 'C0B80000',
-    -5.50: 'C0B00000',
-    -5.25: 'C0A80000',
-    -5.00: 'C0A00000',
-    -4.75: 'C0980000',
-    -4.50: 'C0900000',
-    -4.25: 'C0880000',
-    -4.00: 'C0800000',
-    -3.75: 'C0700000',
-    -3.50: 'C0600000',
-    -3.25: 'C0500000',
-    -3.00: 'C0400000',
-    -2.75: 'C0300000',
-    -2.50: 'C0200000',
-    -2.25: 'C0100000',
-    -2.00: 'C0000000',
-    -1.75: 'BFE00000',
-    -1.50: 'BFC00000',
-    -1.25: 'BFA00000',
-    -1.00: 'BF800000',
-    -0.75: 'BF400000',
-    -0.50: 'BF000000',
-    -0.25: 'BE800000',
-    0.00: '00000000',
-    0.25: '3E800000',
-    0.50: '3F000000',
-    0.75: '3F400000',
-    1.00: '3F800000',
-    1.25: '3FA00000',
-    1.50: '3FC00000',
-    1.75: '3FE00000',
-    2.00: '40000000',
-    2.25: '40100000',
-    2.50: '40200000',
-    2.75: '40300000',
-    3.00: '40400000',
-    3.25: '40500000',
-    3.50: '40600000',
-    3.75: '40700000',
-    4.00: '40800000',
-    4.25: '40880000',
-    4.50: '40900000',
-    4.75: '40980000',
-    5.00: '40A00000',
-    5.25: '40A80000',
-    5.50: '40B00000',
-    5.75: '40B80000',
-    6.00: '40C00000',
-    6.25: '40C80000',
-    6.50: '40D00000',
-    6.75: '40D80000',
-    7.00: '40E00000',
-    7.25: '40E80000',
-    7.50: '40F00000',
-    7.75: '40F80000',
-    8.00: '41000000',
-    8.25: '41040000',
-    8.50: '41080000',
-    8.75: '410C0000',
-    9.00: '41100000',
-    9.25: '41140000',
-    9.50: '41180000',
-    9.75: '411C0000',
-    10.00: '41200000',
-    10.25: '41240000',
-    10.50: '41280000',
-    10.75: '412C0000',
-    11.00: '41300000',
-    11.25: '41340000',
-    11.50: '41380000',
-    11.75: '413C0000',
-    12.00: '41400000',
-    12.25: '41440000',
-    12.50: '41480000',
-    12.75: '414C0000',
-    13.00: '41500000',
-    13.25: '41540000',
-    13.50: '41580000',
-    13.75: '415C0000',
-    14.00: '41600000',
-    14.25: '41640000',
-    14.50: '41680000',
-    14.75: '416C0000',
-    15.00: '41700000'
-}
-
-bandsRangeRTA = {
-    'Low': (118, 292),
-    'Low Mid': (313, 1090),
-    'High Mid': (1170, 5000),
-    'High': (5360, 18660)
-}
